Moved moe router into dedicated variations file
